Created hash.cpp and database.cpp
hash.cpp hashes a file using linux's sha256sum bash command. this will be used for creating a table of file hashes for our project and the code will be reused for any other system calls and will eventually be edited for that

database.cpp has functions for connecting to a database and running queries. my login details are hard coded into it right now, change them before running.